+++ app/Http/Controllers/TheaterController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\ReleaseMovies;
 use App\Models\Theater;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 
 class TheaterController extends Controller
@@ -36,10 +37,15 @@ class TheaterController extends Controller
      */
     public function store(Request $request)
     {
-        $newTheater = new Theater();
-        $newTheater->theater_name = $request->theater_name;
-        $newTheater->save();
-        return $newTheater;
+        $exisitingCity = Theater::where("theater_name", $request->theater_name)->first();
+        if ($exisitingCity) {
+            return json_encode(["status" => "error", "message" => "Theater already exists!"]);
+        } else {
+            $newTheater = new Theater();
+            $newTheater->theater_name = $request->theater_name;
+            $newTheater->save();
+            return json_encode(["status"=>"success","message"=>"Theater Successfully Added"]);
+        }
     }
 
     /**
@@ -73,13 +79,19 @@ class TheaterController extends Controller
      */
     public function update(Request $request, Theater $theater,$id)
     {
-        $existingTheater = Theater::find($id);
-        if($existingTheater){
-            $existingTheater->theater_name = $request->theater_name;
-            $existingTheater->save();
-            return $existingTheater;
+        $childData = Theater::join("release_movies", "release_movies.theater_id", "theaters.id")->where("theaters.id", $id)->get()->count();
+        if ($childData > 0) {
+            return json_encode(["status"=>"error","message"=>"you can not update this data after assign movie"]);
+        } else {
+            $existingTheater = Theater::find($id);
+            if($existingTheater){
+                $existingTheater->theater_name = $request->theater_name;
+                $existingTheater->save();
+                return json_encode(["status"=>"success","message"=>"Theater Successfully Updated"]);
+            }
+            return json_encode(["status"=>"error","message"=>"Theater Not Found"]);
         }
-        return "Theater not found";
+
 
     }
 
@@ -91,12 +103,17 @@ class TheaterController extends Controller
      */
     public function destroy(Theater $theater,$id)
     {
-        $existingTheater = Theater::find($id);
-        if($existingTheater){
-            $existingTheater->delete();
-            return "Theater Successfully deleted";
+        $childData = Theater::join("release_movies", "release_movies.theater_id", "theaters.id")->where("theaters.id", $id)->get()->count();
+        if ($childData > 0) {
+            return json_encode(["status"=>"error","message"=>"you can not delete this data after assign movie"]);
+        } else {
+            $existingTheater = Theater::find($id);
+            if($existingTheater){
+                $existingTheater->delete();
+                return json_encode(["status"=>"success","message"=>"Theater Successfully deleted"]);
+            }
+            return json_encode(["status"=>"error","message"=>"Theater Not Found"]);
         }
-        return "Theater Not found";
     }
 
     public function getAllTheaters(){
